Refactor: Update context provider and navbar logic; remove unused files and adjust peer dependency settings
