+++ .npmrc
@@ -1,3 +1,3 @@
 legacy-peer-deps=true
-strict-peer-dependencies=false
+strict-peer-dependencies=true
 auto-install-peers=true
