+++ package.json
@@ -106,5 +106,9 @@
     "crypto-js": "^4.2.0",
     "@clerk/backend": "^0.35.0",
     "@clerk/clerk-react": "^4.30.0"
+  },
+  "config": {
+    "strict-peer-dependencies": true,
+    "auto-install-peers": true
   }
 }
