--- public/data/course_data.json
@@ -1,32 +1,0 @@
-{
-  "course_1": {
-    "title": "Introduction to Programming",
-    "description": "Learn the basics of programming using Python.",
-    "instructor": "John Doe",
-    "image": "https://example.com/images/course1.jpg"
-  },
-  "course_2": {
-    "title": "Web Development Bootcamp",
-    "description": "Become a full-stack web developer with this comprehensive course.",
-    "instructor": "Jane Smith",
-    "image": "https://example.com/images/course2.jpg"
-  },
-  "course_3": {
-    "title": "Data Science with Python",
-    "description": "Master data science techniques and tools using Python.",
-    "instructor": "Alice Johnson",
-    "image": "https://example.com/images/course3.jpg"
-  },
-  "course_4": {
-    "title": "Machine Learning A-Z",
-    "description": "Learn machine learning from scratch with hands-on projects.",
-    "instructor": "Bob Brown",
-    "image": "https://example.com/images/course4.jpg"
-  },
-  "course_5": {
-    "title": "Advanced JavaScript",
-    "description": "Deep dive into advanced JavaScript concepts and techniques.",
-    "instructor": "Charlie Davis",
-    "image": "https://example.com/images/course5.jpg"
-  }
-}
