--- public/data/course_embeddings.json
@@ -1,7 +1,0 @@
-{
-  "course_1": [0.1, 0.2, 0.3, 0.4, 0.5],
-  "course_2": [0.2, 0.3, 0.4, 0.5, 0.6],
-  "course_3": [0.3, 0.4, 0.5, 0.6, 0.7],
-  "course_4": [0.4, 0.5, 0.6, 0.7, 0.8],
-  "course_5": [0.5, 0.6, 0.7, 0.8, 0.9]
-}
